Create LEADERS IN AN ARRAY .PY
LEADER IN AN ARRAY IS CREATED 
Use two loops. The outer loop runs from 0 to size – 1 and one by one pick all elements from left to right. The inner loop compares the picked element to all the elements on its right side. If the picked element is greater than all the elements to its right side, then the picked element is the leader.